identificador-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
python-dateutil==2.8.2
dateparser==1.1.8
